+++ src/lib/components/delegations/DelegationsBody.tsx
@@ -1,12 +1,22 @@
-import { Flex, RadioGroup, Stack } from "@chakra-ui/react";
+import type { TabProps } from "@chakra-ui/react";
+import {
+  Button,
+  TabList,
+  TabPanel,
+  TabPanels,
+  Tabs,
+  useMultiStyleConfig,
+  useTab,
+} from "@chakra-ui/react";
 import { useRouter } from "next/router";
 import { useEffect, useState } from "react";
 
 import { DelegationsTable, UnbondingsTable } from "../table";
-import { trackUseRadio } from "lib/amplitude";
+import { useMobile } from "lib/app-provider";
 import type { Delegation, Option, TokenWithValue, Unbonding } from "lib/types";
 
-import { RadioCard } from "./radio-card";
+import { MultiBondsRadioCard } from "./radio-card/MultiBondsRadioCard";
+import { SingleBondRadioCard } from "./radio-card/SingleBondRadioCard";
 
 interface DelegationsBodyProps {
   totalDelegations: Option<Record<string, TokenWithValue>>;
@@ -18,6 +28,81 @@ interface DelegationsBodyProps {
   isUnbondingsLoading: boolean;
   bondDenoms: TokenWithValue[];
 }
+
+const getPanelStyle = ({ isMobile }: { isMobile: boolean }) => ({
+  background: isMobile ? "transparent" : "gray.900",
+  border: isMobile ? undefined : "1px solid var(--chakra-colors-gray-700)",
+  borderRadius: "8px",
+  p: isMobile ? 0 : 6,
+});
+
+interface DelegationTabProps extends TabProps {
+  value: string;
+  tokens: Option<Record<string, TokenWithValue>>;
+  isLoading: boolean;
+  bondDenoms: TokenWithValue[];
+}
+
+const DelegationTab = ({
+  isDisabled,
+  value,
+  tokens,
+  isLoading,
+  bondDenoms,
+  ...restProps
+}: DelegationTabProps) => {
+  const tabProps = useTab({ ...restProps });
+  const styles = useMultiStyleConfig("Tabs", tabProps);
+
+  return (
+    <Button
+      __css={styles.tab}
+      display="flex"
+      w="full"
+      mb={0}
+      py={3}
+      borderRadius="8px 8px 0px 0px"
+      color="text.main"
+      sx={{
+        "&[aria-selected=true]": {
+          background: "gray.800",
+          border: "1px solid",
+          borderColor: "gray.700",
+          opacity: "100%",
+          borderBottomColor: "gray.800",
+        },
+        "&[aria-selected=false]": {
+          background: "transparent",
+          border: "1px solid",
+          borderColor: "gray.700",
+          opacity: "70%",
+          borderBottomColor: "transparent",
+        },
+      }}
+      isDisabled={isDisabled}
+      _active={{
+        bg: "unset",
+      }}
+      {...tabProps}
+    >
+      {bondDenoms.length === 1 ? (
+        <SingleBondRadioCard
+          value={value}
+          token={
+            tokens ? tokens[bondDenoms[0].denom] ?? bondDenoms[0] : undefined
+          }
+          isLoading={isLoading}
+        />
+      ) : (
+        <MultiBondsRadioCard
+          value={value}
+          tokens={tokens}
+          isLoading={isLoading}
+        />
+      )}
+    </Button>
+  );
+};
 
 export const DelegationsBody = ({
   totalDelegations,
@@ -31,57 +116,54 @@ export const DelegationsBody = ({
 }: DelegationsBodyProps) => {
   // NOTE: set between "Delegated" and "Unbonding"
   const router = useRouter();
-  const [value, setValue] = useState("Delegated");
+  const [, setValue] = useState("Delegated");
 
   useEffect(() => {
     setValue("Delegated");
   }, [router.query.accountAddress]);
 
+  const isMobile = useMobile();
+
   return (
-    <Flex
-      direction="column"
-      gap={8}
-      p={{ base: 4, md: 8 }}
-      borderRadius="8px"
-      bg="gray.900"
-    >
-      <RadioGroup
-        onChange={(newValue) => {
-          trackUseRadio(newValue.toLocaleLowerCase());
-          setValue(newValue);
-        }}
-        value={value}
-        overflowX="scroll"
-      >
-        <Stack direction={{ base: "column", md: "row" }}>
-          <RadioCard
-            value="Delegated"
-            tokens={totalDelegations}
-            isLoading={isDelegationsLoading}
-            bondDenoms={bondDenoms}
-          />
-          <RadioCard
-            value="Unbonding"
-            tokens={totalUnbondings}
-            isLoading={isUnbondingsLoading}
-            bondDenoms={bondDenoms}
-          />
-        </Stack>
-      </RadioGroup>
-      {value === "Delegated" ? (
-        <DelegationsTable
-          delegations={delegations}
-          rewards={rewards}
+    <Tabs isLazy lazyBehavior="keepMounted" mt={6} w="full">
+      <TabList borderBottom="0px solid" gap={2}>
+        <DelegationTab
+          tokens={totalDelegations}
           isLoading={isDelegationsLoading}
-          isSingleBondDenom={bondDenoms.length === 1}
+          bondDenoms={bondDenoms}
+          value="Delegated"
         />
-      ) : (
-        <UnbondingsTable
-          unbondings={unbondings}
+        <DelegationTab
+          value="Unbonding"
+          tokens={totalUnbondings}
           isLoading={isUnbondingsLoading}
-          isSingleBondDenom={bondDenoms.length === 1}
+          bondDenoms={bondDenoms}
         />
-      )}
-    </Flex>
+      </TabList>
+      <TabPanels
+        background="gray.800"
+        border="1px solid"
+        borderColor="gray.700"
+        borderTopColor="transparent"
+        borderRadius="0px 0px 8px 8px"
+        p={4}
+      >
+        <TabPanel {...getPanelStyle({ isMobile })}>
+          <DelegationsTable
+            delegations={delegations}
+            rewards={rewards}
+            isLoading={isDelegationsLoading}
+            isSingleBondDenom={bondDenoms.length === 1}
+          />
+        </TabPanel>
+        <TabPanel {...getPanelStyle({ isMobile })}>
+          <UnbondingsTable
+            unbondings={unbondings}
+            isLoading={isUnbondingsLoading}
+            isSingleBondDenom={bondDenoms.length === 1}
+          />
+        </TabPanel>
+      </TabPanels>
+    </Tabs>
   );
 };