+++ src/lib/components/delegations/index.tsx
@@ -52,7 +52,15 @@ export const DelegationsSection = ({
   }, [onClose, router.query.accountAddress]);
 
   if (isLoading) return <Loading />;
-  if (!stakingParams) return <ErrorFetching dataName="delegation data" />;
+  if (!stakingParams)
+    return (
+      <ErrorFetching
+        dataName="delegation data"
+        withBorder
+        my={2}
+        hasBorderTop={false}
+      />
+    );
 
   const redelegationCount = redelegations?.length ?? 0;
 
@@ -73,6 +81,7 @@ export const DelegationsSection = ({
         transition="all 0.25s ease-in-out"
       >
         <DelegationInfo
+          hasTotalBonded={totalBonded && Object.keys(totalBonded).length > 0}
           totalBondedCard={
             <TotalCard
               title="Total Bonded"